Updated openstack/openstack Project: openstack/tempest  3510f67d1dbf2d87064057d5db376de9add9b040
Validate floating IP in server['addresses'] scenario

While reviewing nova fix I03be8100155d343eb6a4ea9eda3f1498ad3fb4cf
it was noticed that when you delete a floating IP via the Nova
API which is associated to a server, and using neutron, the
server's instance_info_cache is not directly updated, which is what
populates the server.addresses field. nova-network does refresh the
info_cache when deleting an associated floating IP.

This patch adds validation to a scenario test that runs on both
nova-network and neutron to check this behavior.

Depends-On: I8a8ae8cdbe2d9d77e7f1ae94ebdf6e4ad46eaf00

Change-Id: I5f6b9754289a9e53115320132a2f6629d7f637cb
Related-Bug: #1614538