Updated openstack/openstack Project: openstack-infra/project-config  856183069adeb8878dcdceb836614adec43be1b1
Run diskimage-builder trusty tests on master

The gate-dib-dsvm-functests-ubuntu-trusty test has not been running
since the matches to skip ".*-ubuntu-trusty$" on master were put in
with Idba3883443e53567435c36dcc833b5f3988d8670.  Note dib also has a
"gate-dib-dsvm-functests-ubuntu-trusty-extras-nv" job which is the
non-voting portion of the functional tests (for nice-to-have platforms
like gentoo) -- I have tried to account for this (but it is currently
running anyway, because it doesn't match with the extra bit at the
end).

I think this is just an unfortunate case of dib already doing
multi-platform testing and having the node-type at the end of the job
name before we started to introduce Xenial.

Add this back in, as we want to test on trusty at the very least while
nodepool is building on trusty.

Change-Id: I148e24be59aa67d269b687a771468075d178ab72