Updated openstack/openstack Project: openstack/keystone-specs  32efe5a298ebdc8b892e4c4dd56d003f9d277dd5
Revert spec change for Microversions

We decided to defer support for microversions until we have a change
that actually requires this.

Even though the "old style" identity spec has been moved to the
attic, we had updated this to indicate support for microversions.
So as to not cause confusion, this is now reverted.

Change-Id: I20f284fd760c3bf51b3ebfa7bd91dd64750d9685