Updated openstack/openstack Project: openstack/trove  d7fc09a1bc67d72aa22f58baff5379cd33c4f7a5
Fix pylint Error:PGSQL_RECOVERY_CONFIG

we should use pgsql_recovery_config, the method which has now been
renamed to lowercase. The earlier code used a constant (upper-case)
but in the change this one was likely missed.

pgsql_recovery_config is an @property and should be referenced as such.

This has nothing to do with case per-se, just a missing
function/variable.

Closes-Bug: #1624996
Change-Id: Icc9d794b45503340d988101fec710fae76f0b915