Updated openstack/openstack Project: openstack/python-swiftclient  27f9178ec89259553a770349e0192b0981bdb956
Update help text around container ACL syntax

The patch add the information about the syntax of provding read and
write access to the users within a project.

The current status is:
account:* does not work with tempauth
account does not work with keystoneauth

Furthermore, the concept of "account" is fuzzy. It could be a
"project", but even that doesn't distinguish between a project
id and a project name.

Change-Id: I4462230fe05aa4a742e825f5b6312deb93bc1170
Closes-Bug: #1587677