Updated openstack/openstack Project: openstack-infra/project-config  ff177de614dd2ed513b83da7520d2b83d346e611
Do not 'cd' from current workspace dir in horizon test hooks

So we can copy test artifacts to the current workspace dir, which name
otherwise needs to be passed into test hook. This is needed because
the job name (and consequently the workspace name) can change
independently from test hooks defined in horizon, and thus cannot be
hardcoded.

Needed-By: I63e45ee89711b429d0d878303aefeec4b159125a
Change-Id: I28d84235fb51a49492ed9c96794ba74f317f66b6
Partial-Bug: #1625253