Updated openstack/openstack Project: openstack/nova  193cc78bac609fb82c6ef9e48151ff8fd4268aa6
Do not try to backport when db has older object version

Instance extras are stored as serialized objects in the database and
because of this it is possible to get a version back that is older
than what the client requested. This is in itself not a problem, but
the way we do things right now in conductor we end up trying to
backport to a newer version, which raises InvalidTargetVersion.

This change adds a check to make sure we skip backporting if the
requested version is newer than the actual db version as long as the
major version is the same.

Change-Id: I34ac0abd016b72d585f83ae2ce34790751082180
Closes-Bug: #1596119