Updated openstack/openstack Project: openstack/ironic  20e0b6fa1b7d9fbf3b7fb473c2ad5d7185bf9196
Grub2 by default for PXE + UEFI

This patch is changing the default bootloader for PXE + UEFI to Grub2.
Prior to this patch Elilo was the default but, Elilo is no longer
maintained with the last release dating 2014.

Ironic should have sane defaults therefore we should not rely on an
unmaintained project.

Change-Id: Ib34fcbd37e4698bf272582c96b1540b511b5a892