Updated openstack/openstack Project: openstack/tempest  87d50c41c84268758d5b324b5d5d137429ec747a
Rename renos of 12.2.0

We have released a lot of features via Tempest versions and created a
lot of renos. That makes us difficult to select suitable reno files
when adding new features because of many versions.
So it is nice to rename reno files to add versions as the prefix.
This patch renames renos of 12.2.0 as a first step.

Change-Id: I36523c1758533cca97abad4a28f8c79b022803f0