Updated openstack/openstack Project: openstack/api-site  7c6a1d42453ebb678555b30aaccd12df1d9a9ca6
Add marker file

Add marker files in the root of our tree.

This is needed for publishing in the future:
This is so that we can create a list of directories to ignore when we
rsync documentation builds to a publishing site. Some builds are
published as subdirectories underneath the results of superior
builds.

This allows those superior builds to be rsynced without removing the
other builds that were published under them.

Change-Id: I96a67f022f9fef0794a8a25b6e397177e30d1f5f