Updated openstack/openstack Project: openstack-infra/tripleo-ci  3cfb5c442bc497b68d14ed4d1996cce0727db486
Don't send metrics to graphite in failed jobs

The metrics from a failed job are likely bogus anyway and just
create noise in the Graphite data.  Let's only store the runtimes
from successful jobs.

Because we do some funny business in the trap, we need to first
store the previous exit value and pass it along to postci.
Otherwise by the time we get to postci the exit value we care
about is long gone.

Change-Id: I55db493ce5a1cdc38185d1f805adf4e89e5cbb9c