Updated openstack/openstack Project: openstack/nova  bc1193f60f58fa011f9fcb859017d4f67fdcee67
XenAPI: add unit test for plugin bandwidth

Add unit test for plugin bandwidth. And there is a bug in
the plugin which does not close the opened file. This change
will try to fix it. It also splits _read_proc_net() into two
functions, so that it's easier to make the UT.

Change-Id: Ie871b1d16cf180e9709e46bb0d67573053d14376