Updated openstack/openstack Project: openstack/nova  014b503a57567366043205a234a111f49d71f977
XenAPI: resolve the fetch_bandwidth failure

When introducing interim bridge for XenAPI, it introduced a interface
which is connected to the interim bridge and it has the name as
vifxxxxxxxx-xx. In the xenapi plugin of bandwidth, it uses
the pattern of "^vif*" to filter out VM's interfaces. Due to the
existing of vifxxxxxxxx-xx, it will get exception when parsing the
interface's name basing on "vif<dom#>.<dev>".
So let's give more restrict pattern to filter VM's vifs e.g. change from
"^vif*" to "^vif<dom#>.<dev#>". May consider to use another interface
name for the interface on interim bridge; but need cover both name
patterns when delete interfaces to avoid upgrade issues which may make
things complicated.

Change-Id: I36b96bbe2bb764f3d60cbeb43aa3dbd9a0a43f61
Closes-Bug: #1597551