Updated openstack/openstack Project: openstack/requirements  0837ed63b670ca07e5589c1b0a52a315e57c29fd
Bump g-r python-glanceclient to 2.5.0

Start Ocata with min requirement of 2.5.0. A few changes have merged
since 2.3.0 and we would like to avoid the negative flag on 2.4.0 and
keep the slate clean.

Release 2.5.0 has been in the gate for a few days and can be
assumed to be safe for this bump.

Change-Id: Ibb5e39369a34ce1d34df34472034d9f6c72fff11