Updated openstack/openstack Project: openstack/neutron  a7e528bc8fe2dc957187fdd7eb135dbed8a762cf
Make DHCP agent use 'revision_number'

Change I445974b0e0dabb762807c6f318b1b44f51b3fe15 updated the
'revision' field to 'revision_number' but it missed the DHCP
agent and subsequently broke it's ability to detect stale updates.

This fixes the name in the agent.

This is marked as a partial for 1622616 because one of the reasons
the agent was frequently updating the DHCP port was in reaction
to stale port update messages for its own port.

Partial-Bug: #1622616
Closes-Bug: #1625867
Change-Id: Id41000127e1084f7ff243f8dc9c399999fbdaab4