Updated openstack/openstack Project: openstack/tempest  2c99c44112514a11ef7df8d69726e4ef56675570
Rename renos of 10.0.0 and 11.0.0

We have released a lot of features via Tempest versions and created a
lot of renos. That makes us difficult to select suitable reno files
when adding new features because of many versions.
So it is nice to rename reno files to add versions as the prefix.
This patch renames renos of 10.0.0 and 11.0.0.

Change-Id: I7ccd961f31dbe705a1fafaf6f81b5edd241c6e4d