Updated openstack/openstack Project: openstack/cinder  b1254fabae4ff53aea873b50cfbc6bd5086327ba
Include OPTIONAL_FIELDS in VolumeType OVO class

Current code has OPTIONAL_FIELDS as part of the VolumeType library
instead of as a class attribute of the VolumeType OVO class.  This is
not how it should be, as there are methods -like conditional update-
that rely on the OPTIONAL_FIELDS being there.

This patch moves the OPTIONAL_FIELDS inside the VolumeType class.

Change-Id: I18404a2f32abc6c9d9b7d932a7acd299c4e708e5