Updated openstack/openstack Project: openstack/keystone  c1f9ec275fc03e9c4fd5a198519bb24469f7c533
Handle the exception from creating request token properly

The status code returned larger than 399 not indicate the
signature is invalid, only an empty body implies.

For other reasons which cause the failure of creating request
token this patch show the detail message, so that give us
some clue on where is incorrect.

Related-Bug: #1616424
Change-Id: Id3f0b806630697436340ab97f328743d7f811a14