Updated openstack/openstack Project: openstack/neutron  4ac02616965fb94515147410ed113ecaf36ebf76
Add basic status transition trunk scenario test

While we wait on a solution to getting 802.1q access inside
of the guest VM, it is helpful to at least have a test that
causes the agents to wire up trunk ports and ensure data model
status transitions happen as expected and that connectivity to
the trunk port itself is not disrupted by subport operations.

This test performs those assertions by booting two VMs with
trunk ports, creating subports on one, asserting statuses,
moving the subports to the other, and asserting statuses again
while ensuring connectivity to the VM in between.

This test uncovered both of the related bugs.

Related-Bug: #1625136
Related-Bug: #1625135
Change-Id: I5f634e3bcc54f5420ba49027b4fb250a13fe277a