Updated openstack/openstack Project: openstack/python-openstackclient  69c4f605ecce597af3a574d2f7c27c1dbaa17989
Fix "volume unset" command pass normally when nothing specified

When nothing specified in "volume unset" command, there
will be an error message says that the "--properties"
option is required, it is unusual behaviour, this patch
fix it and also add unit test for it.
Also, this patch add unit test for "volume show" command
by the way.

Change-Id: I5b5d587670acf0af4262b8521292455bf9f60fe5
Partial-bug: #1588588