Updated openstack/openstack Project: openstack/python-keystoneclient  dc9e4491779c0347f4eb770de03827b967541bf3
Fix non-ascii attributes

It's possible to set and get custom attributes through the API, and
it's also possible to use any Unicode strings, not only ASCII range.
It works perfectly fine when accessing the API directly using
e.g. curl.

However, keystoneclient stumbles on the non-ascii keys because
it tries to set this as an attribute for the resource class and fails
because Python 2.7 does not support non-ascii identifiers:

https://docs.python.org/2.7/reference/lexical_analysis.html#identifiers

So change the logic to skip setting non-ascii attributes; they are still
available in the dict representation.

Closes-Bug: #1626403
Change-Id: I267188cdb1d303e3d0fb6bd3229b606f4fe9b2d8