Updated openstack/openstack Project: openstack/nova  f55ea7868059ae2b3c54ee96f708ca4480226f7c
Fix database poison warnings, part 10

The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes all the warnings from:

    unit.compute.test_compute_mgr.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I72da6b07a854a69c275258027afa73e7bfaeba17
Related-Bug: #1568414