Updated openstack/openstack Project: openstack/nova  832e2771983508e943fb2b12be481e31a38197e6
Fix database poison warnings, part 13

The following warning appears in the unit test logs a number of times.

    "UserWarning: This test uses methods that set internal oslo_db
state, but it does not claim to use the database. This will conflict
with the setup of tests that do use the database and cause failures
later."

This patch fixes the warnings from:

    unit.virt.vmwareapi.test_driver_api.py

Note that this warning is only emitted once per unit test worker, so new
offenders will show up in the logs each time you fix a test until they
are all gone.

Change-Id: I2dd21ec8c970161f695b2f37ae0973343678a0c4
Related-Bug: #1568414