Updated openstack/openstack Project: openstack/nova  a4d846f43962d5a2f53bc3d51e36a3e89115e458
[placement] Stringify class and provider uuid in error

When attempting to make an allocation of a particular class of
resource against a resource provider which has no inventory for that
class, an InvalidInventory exception is raised. This was workign
okay but the associated message was displaying stringified sets
of resource class ids, and not the resource classes names.

This changes joins the sets into strings for both resource classes
and provider uuids and turns the class indexes into their names.
A functional test which was checking for the exception has been
updated to also check the exception's message.

Change-Id: Ife38220da1069ffb6da26a4f8e3b954f0dc12f13
Closes-Bug: #1620748