Updated openstack/openstack Project: openstack/cinder  860c183e823ec3622c0d9b08b0273f9e5fd94c36
Dell SC: Error attaching after LV-AFO

Fixed miscellaneous errors dealing with the state of Dell's
Live Volumes after a LV AFO event.

Was looking for the original volume and sc server object
after the primary was down. We now look for the new primary
and attach that.

Ensure export was failing to properly kick off a swap
back to the original primary if it was available.

Fixed an error with deleting replications.

Added a workaround to finding a ScLiveVolume object. Previous
method could return stale information.

Change-Id: I58889adbcf279d4b0107cbc51e1dace888a3fb15
Closes-Bug: #1617401