Updated openstack/openstack Project: openstack/nova  e7601e5c1c5ce7c5284f1c98c0bb05e3a76bd9f8
Work around tests that don't use nova.test as a base

The db migration tests use a bunch of setup inside oslo_db that is
only exported as a base test class. This means that none of these do
the standard setup that the rest of the Nova tests do, including
things like logging capture.

Long term, the relationship between nova and oslo_db needs to be fixed
so that these are available as fixtures, not just base class. However,
in the short term we can setup logging correctly before initializing
the rest of the test class to keep these from scrambling the test
output.

Change-Id: I3c663d276cffa6146ac8f45bde7d70f7aece0c89