Updated openstack/openstack Project: openstack/openstack-manuals  0cf6cde4619c33321ce8d14e04a546e2392580da
don't hyphenate "highly {available,scalable}"

The majority of occurrences of "highly available" in the manuals are not
hyphenated, but a minority are, which is inconsistent.  Similarly with
"highly scalable".

According to

  https://www.quora.com/Should-highly-respected-be-hyphenated-Is-there-a-general-rule-about-hyphenation

the rule is that words should be hyphenated only to avoid confusion,
but in the case of "highly available OpenStack" or similar, it is
clear that "highly" modifies "available" rather than "OpenStack",
because "highly OpenStack" makes no sense.

Therefore we remove the hyphens from the minority and go with the
majority.

Change-Id: Ifdd029b1c6bfbfc1f912edd8ed04f6c1148e36d0