Updated openstack/openstack Project: openstack/nova  f46a662c56e037af5f18d7223b7f3268b0eb4a89
Remove PCI parent_addr online migration

This migration was added in mitaka and should have been done either in
mitaka or newton. Newton had some migrations that came after it which
were higher-visibility and should have forced everyone to have done
this by now.

Because of the nature of the migration, there isn't really an easy or
efficient way to validate that they have done this like we have done
for some other migrations.

Change-Id: Idb033e9e52b149372308eabb19c5774e10c56156