Updated openstack/openstack Project: openstack-infra/tripleo-ci  67b94b7e00f67a812812af2283c22f292b1d13dd
Use low-memory-usage.yaml in ci

Now that we have an environment in t-h-t that is designed to do the
same thing, we should use it instead of the ci-specific version.

Because the ci file had extraconfig as well, we have to leave it
for now, but we should probably move those settings to the t-h-t
version of the file as well.

Change-Id: I9d3ec8b698ca2ee985334f40ad8cd495c90db9b7