Updated openstack/openstack Project: openstack-infra/project-config  4514053e89d923d7cc36f68775417149053818e0
Add searchlight-ui-core to searchlight-ui ACL

We have contributors to searchlight-ui who are not eligible to be core
developers either on searchlight or horizon but who we would like to
be able to approve changes to searchlight-ui.

This will require a new group searchlight-ui-core; I'm not sure if that
happens automatically or is a manual process. I can be the initial
member.

Change-Id: I06c36a817a1aebecca1a587a17088ced2beb1ae2