Updated openstack/openstack Project: openstack-infra/project-config  56cb600f21768da580b5d30634f1d81940901f5e
Demote OSA multi-region swift job from voting to non-voting

We've discovered some issues with the check which are holding back
the progress of out update to ensure that the repository is tested
using Ansible 2.1.x. It's been decided that the job should be
demoted as the multi-region test is less important for now.

Change-Id: Ie77dbe9dc3f432bc20c608b68e0e7fceb1e7d5c1