Updated openstack/openstack Project: openstack/tripleo-heat-templates  56057325a95bb521e2ed8a322223c1f858ab25fe
Make apache-based services use network-dependent servername

Currently the servername is incorrectly set for the services running
over apache. It currently takes the default value which is just the
regular FQDN, when the services actually might be running on
different IPs that require alternative FQDNs.

This fixes that by filling that value from a fact in hiera that's
dependant on the service's network.

Closes-Bug: #1625677
Change-Id: Ib7ea5fd2d18a376eaa2f5a3fa5687cb9b719a8e2