Updated openstack/openstack Project: openstack/nova  0ebb0b019e5b5d336c500df3baa9099df39b4638
Remove nova.image.s3 and configs

This code has no unit tests, and all of its config options were
deprecated in Newton. As pointed out by Matt Riedemann, it's time for
this code to be removed if nobody has stepped up to make sure it is well
tested and functional. This also removes the last dependency on boto.

Change-Id: I60a304f242ce342d16b6465559b04194041ff240