Updated openstack/openstack Project: openstack/trove  7767bb6fc717aec6a928661435664af583deb5ea
update contributing.rst to reflect release notes and bugs

Hitherto we have tried to enforce the requirement that changes have
Launchpad bugs. More openstack projects have moved away from this
requirement and it may be time for Trove to do the same.

We use reno for release notes and therefore if a release note is not
specified in the commit, it will not make it to the official release
notes page. We should therefore require release notes for all user
visible changes.

Change-Id: If72fc47c1d699b20a85574f5945ecff35cc0d008