Updated openstack/openstack Project: openstack/neutron  bf39b0b35defc4021f55eb5497a628b2fcc1909c
tests: catch eventlet.Timeout exception

This exception kills the running test worker with all test cases
scheduled to it, probably because unittest module is not capable of
surviving it.

This patch makes all test cases to catch the exception and convert it to
unittest' friendly failure mode (triggered with self.fail).

Change-Id: I5b0d1efa458ca57dfce637dc75d419fe127751ed
Closes-Bug: #1625221