Updated openstack/openstack Project: openstack/nova  fa856ab21b5db21901cccae555bc4e11d6535e25
[placement] Adjust the name of the gabbi tests

Recently there was some confusion caused by the fact the gabbi tests
used by the placement API have names like this:

    gabbi.suitemaker.test_placement_api_usage_create_provider

This means that the names cannot be matched against tox interactions
like:

    tox -efunctional -- nova.tests.functional.api.openstack

This change makes it so the tests are generated with names like:

    nova.tests.functional.api.openstack.placement.test_placement_api.usage_create_provider

So this change is submitted for review if people care. The other
option is to:

    tox -efunctional placement

gabbi was updated to allow this functionality in a clean form, so
this depends on a change to global requirements but can work
without it if required.

Depends-On: Ie93c0e680a80fc7f157e2b5b9dfc2ab17f934f10
Change-Id: I8579bf7529a7298c4d049e47cc3cde1731a5e666