Updated openstack/openstack Project: openstack-infra/project-config  41aebc055484274ed59cc939866bcea76207a344
Cleanup partial translated files

Raise threshold on when we remove partially translated files from 20 per
cent to 40 per cent. New files that get added need to be translated by
75 per cent.

We agreed in the past that files with less than 60 per cent
can be improved but let's not go that high directly.

Change-Id: I2c9a8c73d8364124b98ae8fcc0c0eb5221507d46